Comment-out sleep, see how CI works.
